Create README.md
A project of mine when I was in college in Computer Assembly Subject. The game is based on Deal or No Deal game. A user can choose briefcases and the banker will offer base on the briefcases amount that is still available which ask the user if it's Deal or No Deal.